Added authorization

+++ routes/post.js
@@ -66,6 +66,7 @@ router.post('', checkAuth, multer({ storage: storage }).single('image'), (req, r
     title: req.body.title,
     content: req.body.content,
     imagePath: url + '/images/' + req.file.filename,
+    creator: req.userData.userId,
   });
   post.save().then((createdPost) => {
     res.status(201).json({
@@ -89,15 +90,24 @@ router.put('/:id', checkAuth, multer({ storage: storage }).single('image'), (req
     title: req.body.title,
     content: req.body.content,
     imagePath: imagePath,
+    creator: req.userData.userId,
   });
-  Post.updateOne({ _id: req.params.id }, post).then((result) => {
-    res.status(200).json({ message: 'Update successful!' });
+  Post.updateOne({ _id: req.params.id, creator: req.userData.userId }, post).then((result) => {
+    if (result.nModified > 0) {
+      res.status(200).json({ message: 'Update successful!' });
+    } else {
+      res.status(401).json({ message: 'Not authorized!' });
+    }
   });
 });
 
 router.delete('/:id', checkAuth, (req, res, next) => {
-  Post.deleteOne({ _id: req.params.id }).then((result) => {
-    res.status(200).json({ message: 'Post deleted!' });
+  Post.deleteOne({ _id: req.params.id, creator: req.userData.userId }).then((result) => {
+    if (result.n > 0) {
+      res.status(200).json({ message: 'Post deleted!' });
+    } else {
+      res.status(401).json({ message: 'Not authorized!' });
+    }
   });
 });
